fix: input type fixes
