+++ .npmrc
@@ -1,2 +1,1 @@
-force = true
 access = public
