+++ packages/core/src/components/question/input-types/email/EmailInput.tsx
@@ -1,15 +1,11 @@
 "use client";
-import { InputComponentType, registerInputComponent } from "@eavfw/quickform-core";
+ 
 import { BaseInputComponent } from "../baseinput/BaseInputComponent";
-import { emailInputSchema } from "./EmailInputSchema";
-import { IconType } from "../../../icons/IconResolver";
+import { emailInputSchema } from "./EmailInputSchema"; 
+import { InputComponentType, registerInputComponent } from "../../../../services/defaults/DefaultInputTypeResolver";
+import { EmailProperties } from "../../../../model/InputType";
 
-export type EmailProperties = {
-    inputType: "email";
-    defaultValue?: string;
-    beforeIcon?: IconType;
-    afterIcon?: IconType
-}
+
 
 export const EmailInput: InputComponentType<EmailProperties> = (props) => {
     return <BaseInputComponent type="email" {...props} />
